requirements.txt
asgiref==3.5.2
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-summernote==0.8.20.0
django-webpush==0.3.4
http-ece==1.1.0
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.5
py-vapid==1.9.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pywebpush==1.9.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
webencodings==0.5.1
